Add set pass, set adm, leave Group, set onlyInvite And tested new features

+++ back/src/chat/chat.service.ts
@@ -1,9 +1,10 @@
 import { BadRequestException, Injectable } from "@nestjs/common";
 import { PrismaService } from "src/prisma/prisma.service";
 import { messageToClient, messageToServer } from "./dto/chat.interface";
-import { CreateGroupDto, GroupActionsDto, InviteToGroupDto, PassowordChannel, SetAdm } from "./dto/chat.dto";
+import { CreateGroupDto, GroupActionsDto, InviteToGroupDto, PassowordChannel, SetAdm, SetOnlyInvite } from "./dto/chat.dto";
 import { GroupService } from "./services/group.service";
 import * as argon from 'argon2';
+import { Group, User } from "@prisma/client";
 
 //TODO later -- NEED IMPLEMENT A FRONT END ROOMS AND MESSAGES TO ROOMS JUST SIMPLE
 @Injectable()
@@ -78,7 +79,6 @@ export class ChatService {
     /**
     * @brief when create a group in controller the front end should use this socket to join the owner socket in group
     */
-    //TODO 2024 1- precisa verificar se ja esta criando a role de adm e a partir disso criar comandos para certas acoes
     async joinOwnerInGroup(groupActionsDto: GroupActionsDto): Promise<boolean> {
         const group = await this.groupService.getGroupByName(groupActionsDto.groupName,);
         const target = await this.groupService.getUserByUsername(groupActionsDto.username,);
@@ -114,7 +114,6 @@ export class ChatService {
         })
     }
 
-    // TEST
     async setUserAsAdm(setAdm: SetAdm){
         const group = await this.groupService.getGroupByName(setAdm.groupName);
         const adm = await this.groupService.getUserByUsername(setAdm.admUsername);
@@ -140,6 +139,7 @@ export class ChatService {
     }
 
     // TEST 1.2 talves nao precise disso
+    //TODO FIX se continuar para nao deletar quem nao pode owner e adm
     async removeAdm(setAdm: SetAdm){
         const group = await this.groupService.getGroupByName(setAdm.groupName);
         const adm = await this.groupService.getUserByUsername(setAdm.admUsername);
@@ -150,11 +150,14 @@ export class ChatService {
         const isMember = await this.groupService.isMemberInGroup(user.id, group.id);
 
         if (!isMember)
-            throw new BadRequestException(`User ${setAdm.userToBeAdm} in not a member in group`)
+            throw new BadRequestException(`${setAdm.userToBeAdm} in not a member in group`)
         if (!isUserAdm)
-            throw new BadRequestException(`User ${setAdm.userToBeAdm} is not a adm in Group`)
+            throw new BadRequestException(`${setAdm.userToBeAdm} is not a adm in Group`)
         if (!isAdm)
-            throw new BadRequestException(`User ${setAdm.admUsername} is not a adm in Group`)
+            throw new BadRequestException(`${setAdm.admUsername} is not a adm in Group`)
+
+        if(group.ownerId == user.id)
+            throw new BadRequestException(`${user.user} the owner can't be removed the from adm list`)
 
         await this.prisma.groupAdmin.deleteMany({
             where: {
@@ -172,7 +175,6 @@ export class ChatService {
     //convidar para jogar
     //acessar perfil
 
-    //TEST com e sem senha
     async changeChannelPass(passwordChannel: PassowordChannel){
         const owner = await this.groupService.getUserByUsername(passwordChannel.ownerUsername);
         const group = await this.groupService.getGroupByName(passwordChannel.groupName);
@@ -181,15 +183,93 @@ export class ChatService {
             throw new BadRequestException(`user ${owner.user} is not the owner in channel ${group.name}`);
 
         if (passwordChannel.password){
+            const password = await argon.hash(passwordChannel.password);
             await this.prisma.group.update({
                 where: { id:group.id},
-                data: {password: passwordChannel.password}
+                data: {password: password, type: "PROTECT"}
             });
         } else {
             await this.prisma.group.update({
                 where: {id: group.id},
-                data: { password: null},
+                data: { password: null, type: "PUBLIC"},
             });
         }
+    }
+
+    async setChannelOnlyInvite(setOnlyInvite : SetOnlyInvite) {
+        const owner = await this.groupService.getUserByUsername(setOnlyInvite.ownerUsername);
+        const group = await this.groupService.getGroupByName(setOnlyInvite.groupName);
+        if (group.ownerId !== owner.id)
+            throw new BadRequestException(`user ${owner.user} is not the owner in channel ${group.name}`);
+        await this.prisma.group.update({
+            where: { id:group.id},
+            data: {password: null, type: "PRIVATE"}
+        });
+    }
+
+    async leaveChannel(username: string, groupName: string) {
+        const user = await this.groupService.getUserByUsername(username);
+        const group = await this.groupService.getGroupByName(groupName);
+
+        const isMember = await this.groupService.isMemberInGroup(user.id, group.id);
+        if (!isMember)
+            throw new BadRequestException(`${user.user} is not a member in ${group.name}`)
+        if (group.ownerId == user.id){
+            var emptyGroup = await this.passOwner(user, group);
+        }
+
+        await this.prisma.groupMembership.deleteMany({
+            where: {
+                userId: user.id,
+                groupId: group.id,
+            },
+        });
+        await this.prisma.groupAdmin.deleteMany({
+            where: {
+                userId: user.id,
+                groupId: group.id,
+            },
+        });
+        if (emptyGroup){
+            await this.prisma.group.delete({ where: { id: group.id } });
+        }
+    }
+
+    async passOwner(user: User, group: Group): Promise<boolean> {
+        const admToOwner = await this.prisma.groupAdmin.findFirst({
+            where: {
+                groupId: group.id,
+                NOT: {userId: user.id,},
+            },
+        });
+        if (admToOwner){
+            await this.prisma.group.update({
+                where: { id: group.id,},
+                data: { ownerId: admToOwner.userId,},
+            });
+        } else {
+            const otherMember = await this.prisma.groupMembership.findFirst({
+                where: {
+                    groupId: group.id,
+                    NOT: { userId: user.id, },
+                },
+            });
+
+            if (otherMember) {
+                await this.prisma.group.update({
+                    where: { id: group.id },
+                    data: { ownerId: otherMember.userId },
+                });
+                await this.prisma.groupAdmin.create({
+                    data: {
+                        userId: otherMember.userId,
+                        groupId: group.id,
+                    },
+                });
+            } else {
+                return true;
+            }
+        }
+        return false;
     }
 }
